Split catalog API methods to ServiceCatalogApi
